+++ src/customwidgets/actionbuttondelegate.h
@@ -1,0 +1,33 @@
+#ifndef ACTIONBUTTONDELEGATE_H
+#define ACTIONBUTTONDELEGATE_H
+
+#include <QStyledItemDelegate>
+
+class ActionButtonDelegate : public QStyledItemDelegate
+{
+    Q_OBJECT
+public:
+    explicit ActionButtonDelegate(QObject* parent = nullptr);
+
+    void paint(QPainter* painter, const QStyleOptionViewItem& option, const QModelIndex& index) const override;
+
+    QSize sizeHint(const QStyleOptionViewItem& option, const QModelIndex& index) const override;
+
+    void setCharacter(QChar character);
+
+    bool editorEvent(QEvent* event,
+                     QAbstractItemModel* model,
+                     const QStyleOptionViewItem& option,
+                     const QModelIndex& index) override;
+
+signals:
+    void clicked(const QModelIndex& index);
+
+private:
+    QRect buttonRect(const QStyleOptionViewItem& option) const;
+    bool isButtonEnabled(const QModelIndex& index) const;
+
+    QChar _character;
+};
+
+#endif // ACTIONBUTTONDELEGATE_H