+++ src/dialogs/importmbcdialog.h
@@ -4,11 +4,14 @@
 #include <QDialog>
 
 #include "customwidgets/centeredbox.h"
+#include "models/mbcregistermodel.h"
 
 /* Forward declaration */
+class ActionButtonDelegate;
 class GuiModel;
-class MbcRegisterModel;
+class GraphDataModel;
 class MbcRegisterFilter;
+class MbcUpdateModel;
 
 namespace Ui {
 class ImportMbcDialog;
@@ -19,7 +22,7 @@ class ImportMbcDialog : public QDialog
     Q_OBJECT
 
 public:
-    explicit ImportMbcDialog(GuiModel* pGuiModel, MbcRegisterModel* pMbcRegisterModel, QWidget* parent = nullptr);
+    explicit ImportMbcDialog(GuiModel* pGuiModel, GraphDataModel* _pGraphDatamodel, QWidget* parent = nullptr);
     ~ImportMbcDialog();
 
 public slots:
@@ -28,6 +31,7 @@ public slots:
 private slots:
     void updateTextFilter();
     void selectMbcFile();
+    void importSelectedRegisters();
     void visibleItemsDataChanged();
     void registerDataChanged();
     void handleSelectAllClicked(Qt::CheckState state);
@@ -35,13 +39,17 @@ private slots:
 private:
     void setSelectedSelectionstate(Qt::CheckState state);
     void updateMbcRegisters(QString filePath);
+    void handleAcceptUpdate(const QModelIndex& index);
 
     Ui::ImportMbcDialog *_pUi;
 
     CenteredBoxProxyStyle _centeredBoxStyle;
+    std::unique_ptr<ActionButtonDelegate> _pUpdateDelegate;
 
     GuiModel * _pGuiModel;
-    MbcRegisterModel * _pMbcRegisterModel;
+    GraphDataModel* _pGraphDataModel;
+    MbcRegisterModel _mbcRegisterModel;
+    MbcUpdateModel* _pMbcUpdateModel;
 
     MbcRegisterFilter * _pTabProxyFilter;
 };