+++ tests/communication/tst_communication.cpp
@@ -10,7 +10,8 @@
 #include "testslavedata.h"
 
 #include <QMap>
-#include <QtTest/QtTest>
+#include <QSignalSpy>
+#include <QTest>
 
 using State = ResultState::State;