+++ tests/communication/tst_modbusmaster.cpp
@@ -7,7 +7,8 @@
 
 #include <QMap>
 #include <QMetaType>
-#include <QtTest/QtTest>
+#include <QSignalSpy>
+#include <QTest>
 
 Q_DECLARE_METATYPE(Result<quint16>);